requirements.txt
asgiref==3.11.1
Django==6.0.2
django-cors-headers==4.9.0
djangorestframework==3.16.1
gunicorn==25.0.1
packaging==26.0
sqlparse==0.5.5
yt-dlp==2026.2.4
